[#157343689] Rename package com.trongntd.themovie to com.ea.themovie
