+++ app/src/androidTest/java/com/ea/themovie/RecyclerViewItemCountAssertion.java
@@ -1,0 +1,36 @@
+package com.ea.themovie;
+
+import android.support.test.espresso.NoMatchingViewException;
+import android.support.test.espresso.ViewAssertion;
+import android.support.v7.widget.RecyclerView;
+import android.view.View;
+
+import org.hamcrest.Matcher;
+
+import static android.support.test.espresso.matcher.ViewMatchers.assertThat;
+import static org.hamcrest.Matchers.is;
+
+public class RecyclerViewItemCountAssertion implements ViewAssertion {
+
+    private final Matcher<Integer> matcher;
+
+    public RecyclerViewItemCountAssertion(int expectedCount) {
+        this.matcher = is(expectedCount);
+    }
+
+    public RecyclerViewItemCountAssertion(Matcher<Integer> matcher) {
+        this.matcher = matcher;
+    }
+
+    @Override
+    public void check(View view, NoMatchingViewException noViewFoundException) {
+        if (noViewFoundException != null) {
+            throw noViewFoundException;
+        }
+
+        RecyclerView recyclerView = (RecyclerView) view;
+        RecyclerView.Adapter adapter = recyclerView.getAdapter();
+        assertThat(adapter.getItemCount(), matcher);
+    }
+
+}
