+++ app/src/main/java/com/ea/themovie/api/Error.java
@@ -1,0 +1,5 @@
+package com.ea.themovie.api;
+
+public class Error {
+    public static final int UNKNOWN = 1;
+}
