+++ app/src/main/java/com/ea/themovie/entity/MovieList.java
@@ -1,0 +1,10 @@
+package com.ea.themovie.entity;
+
+import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+public class MovieList {
+    @SerializedName("results")
+    public List<Movie> movies;
+}
