+++ app/src/main/java/com/ea/themovie/entity/MovieVideo.java
@@ -1,0 +1,14 @@
+package com.ea.themovie.entity;
+
+import com.google.gson.annotations.SerializedName;
+
+import java.util.List;
+
+public class MovieVideo {
+
+    @SerializedName("id")
+    public String movieId;
+
+    @SerializedName("results")
+    public List<Video> videos;
+}
