+++ app/src/main/java/com/ea/themovie/entity/Video.java
@@ -1,0 +1,11 @@
+package com.ea.themovie.entity;
+
+public class Video {
+    public String key;
+
+    public String name;
+
+    public String site;
+
+    public String type;
+}
