+++ app/src/main/java/com/ea/themovie/fragment/BaseFragment.java
@@ -1,0 +1,38 @@
+package com.ea.themovie.fragment;
+
+import android.content.Context;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentActivity;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+
+public class BaseFragment extends Fragment{
+    protected int layoutResId;
+
+    protected View rootView;
+    protected FragmentActivity activity;
+
+    @Nullable
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        if (layoutResId > 0) {
+            Log.e("vo day", "vo day");
+            rootView = inflater.inflate(layoutResId, container, false);
+        } else {
+            rootView = super.onCreateView(inflater, container, savedInstanceState);
+        }
+        return rootView;
+    }
+
+    @Override
+    public void onAttach(Context context) {
+        super.onAttach(context);
+        activity = (FragmentActivity) context;
+    }
+}
