+++ app/src/main/java/com/ea/themovie/fragment/MostRatedMovieFragment.java
@@ -1,0 +1,8 @@
+package com.ea.themovie.fragment;
+
+public class MostRatedMovieFragment extends ListMovieFragment {
+    @Override
+    public void loadData() {
+        listMoviePresenter.loadMostRatedMovies(1);
+    }
+}
