+++ app/src/main/java/com/ea/themovie/fragment/PopularMovieFragment.java
@@ -1,0 +1,8 @@
+package com.ea.themovie.fragment;
+
+public class PopularMovieFragment extends ListMovieFragment {
+    @Override
+    public void loadData() {
+        listMoviePresenter.loadPopularMovies(1);
+    }
+}
