--- app/src/main/java/com/trongntd/themovie/MainActivity.java
@@ -1,28 +1,0 @@
-package com.trongntd.themovie;
-
-import android.support.design.widget.TabLayout;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-import com.trongntd.themovie.adapter.HomePagerAdapter;
-import com.trongntd.themovie.entity.Movie;
-import com.trongntd.themovie.presenter.ListMoviePresenter;
-
-import java.util.List;
-
-public class MainActivity extends AppCompatActivity{
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        ViewPager viewPager = findViewById(R.id.vp_home_page);
-        HomePagerAdapter adapter = new HomePagerAdapter(this, getSupportFragmentManager());
-        viewPager.setAdapter(adapter);
-
-        TabLayout tabLayout = findViewById(R.id.view_pager_tab);
-        tabLayout.setupWithViewPager(viewPager);
-    }
-
-}
