--- app/src/main/java/com/trongntd/themovie/api/Error.java
@@ -1,5 +1,0 @@
-package com.trongntd.themovie.api;
-
-public class Error {
-    public static final int UNKNOWN = 1;
-}
