--- app/src/main/java/com/trongntd/themovie/entity/Movie.java
@@ -1,22 +1,0 @@
-package com.trongntd.themovie.entity;
-
-import com.google.gson.annotations.SerializedName;
-
-public class Movie {
-    public int id;
-
-    public String title;
-
-    public String overview;
-
-    @SerializedName("poster_path")
-    public String posterPath;
-
-    @SerializedName("release_date")
-    public String releaseDate;
-
-    @SerializedName("vote_average")
-    public String voteAverage;
-
-    public String isFavorited;
-}
