--- app/src/main/java/com/trongntd/themovie/entity/MovieList.java
@@ -1,10 +1,0 @@
-package com.trongntd.themovie.entity;
-
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class MovieList {
-    @SerializedName("results")
-    public List<Movie> movies;
-}
