--- app/src/main/java/com/trongntd/themovie/entity/MovieVideo.java
@@ -1,14 +1,0 @@
-package com.trongntd.themovie.entity;
-
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class MovieVideo {
-
-    @SerializedName("id")
-    public String movieId;
-
-    @SerializedName("results")
-    public List<Video> videos;
-}
