--- app/src/main/java/com/trongntd/themovie/entity/Video.java
@@ -1,11 +1,0 @@
-package com.trongntd.themovie.entity;
-
-public class Video {
-    public String key;
-
-    public String name;
-
-    public String site;
-
-    public String type;
-}
