--- app/src/main/java/com/trongntd/themovie/fragment/MostRatedMovieFragment.java
@@ -1,8 +1,0 @@
-package com.trongntd.themovie.fragment;
-
-public class MostRatedMovieFragment extends ListMovieFragment {
-    @Override
-    public void loadData() {
-        listMoviePresenter.loadMostRatedMovies(1);
-    }
-}
