--- app/src/main/java/com/trongntd/themovie/fragment/PopularMovieFragment.java
@@ -1,8 +1,0 @@
-package com.trongntd.themovie.fragment;
-
-public class PopularMovieFragment extends ListMovieFragment {
-    @Override
-    public void loadData() {
-        listMoviePresenter.loadPopularMovies(1);
-    }
-}
