--- app/src/main/java/com/trongntd/themovie/presenter/BasePresenter.java
@@ -1,13 +1,0 @@
-package com.trongntd.themovie.presenter;
-
-public class BasePresenter<V> {
-    protected V view;
-
-    public void attachView(V viewAction){
-        this.view = viewAction;
-    }
-
-    public void detachView(){
-        this.view = null;
-    }
-}
