--- app/src/main/java/com/trongntd/themovie/util/Constants.java
@@ -1,7 +1,0 @@
-package com.trongntd.themovie.util;
-
-public class Constants {
-    public static final String API_DATE_FORMAT = "yyyy-MM-dd";
-    public static final String API_BASE_URL = "https://api.themoviedb.org/";
-    public static final String API_KEY = "af1153f5d5bee65b2d2a6d0c8aa9447f";
-}
